requirements.txt
fastapi[standard]
passlib
pyjwt
aiohttp
asyncio
python-decouple
bcrypt